Move builder classes to a new subfolder
