+++ RyuSocks.Generator/AuthMethodGenerator.cs
@@ -16,6 +16,7 @@
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using RyuSocks.Generator.Builder;
 using System;
 using System.Collections.Immutable;
 using System.Linq;
