+++ RyuSocks.Generator/Builder/CodeBuilder.cs
@@ -1,6 +1,6 @@
 using System.Text;
 
-namespace RyuSocks.Generator
+namespace RyuSocks.Generator.Builder
 {
     // Original source: https://github.com/Ryujinx/Ryujinx/blob/1df6c07f78c4c3b8c7fc679d7466f79a10c2d496/src/Ryujinx.Horizon.Generators/CodeGenerator.cs
     class CodeBuilder : AbstractBuilder
